third_party/blink/web_tests/external/wpt/screen-capture/permissions-policy-video.https.sub-expected.txt

This is a testharness.js-based test.
PASS Default "display-capture" permissions policy ["self"] allows the top-level document.
PASS Default "display-capture" permissions policy ["self"] allows same-origin iframes.
FAIL Default "display-capture" permissions policy ["self"] disallows cross-origin iframes. assert_equals: Failed to execute 'getDisplayMedia' on 'MediaDevices': Access to the feature "display-capture" is disallowed by permission policy.. expected "NotAllowedError" but got "SecurityError"
PASS permissions policy "display-capture" can be enabled in cross-origin iframes using "allow" attribute.
FAIL permissions policy "display-capture" can be disabled in same-origin iframes using "allow" attribute. assert_equals: Failed to execute 'getDisplayMedia' on 'MediaDevices': Access to the feature "display-capture" is disallowed by permission policy.. expected "NotAllowedError" but got "SecurityError"
Harness: the test ran to completion.
